🚚rename: 리워드 페이지 폴더명 수정
