+++ frontend/src/routes/rewardPageF/companyInfoF/CompanyInfo.jsx
