+++ frontend/src/routes/rewardPageF/tutorialF/receiptTutoF/ReceiptTutorial.jsx
