+++ frontend/src/routes/rewardPageF/tutorialF/receiptTutoF/ReceiptTutorial.module.css
