+++ frontend/src/routes/rewardPageF/tutorialF/videoTutoF/VideoTutorial.jsx
